separated kmergenerator to separate folder
